rename: log -> savelog로 변경
