+++ app/src/main/res/layout/fragment_savelog.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.treasurehunt.ui.login.LogViewModel" />
+            type="com.treasurehunt.ui.savelog.SaveLogViewModel" />
     </data>
 
     <androidx.core.widget.NestedScrollView
@@ -34,7 +34,7 @@
 
             <TextView
                 android:id="@+id/tv_title"
-                style="@style/LogTitle"
+                style="@style/SaveLogTitle"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="40dp"
@@ -58,7 +58,7 @@
 
             <TextView
                 android:id="@+id/tv_photo"
-                style="@style/LogPhoto"
+                style="@style/SaveLogPhoto"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="16dp"
@@ -89,11 +89,11 @@
                 app:layout_constraintStart_toEndOf="@+id/ib_select_photo"
                 app:layout_constraintTop_toTopOf="@+id/ib_select_photo"
                 app:dataset="@{viewModel.images}"
-                tools:listitem="@layout/item_log" />
+                tools:listitem="@layout/item_savelog" />
 
             <TextView
                 android:id="@+id/tv_text"
-                style="@style/LogPhoto"
+                style="@style/SaveLogPhoto"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="16dp"
@@ -104,7 +104,7 @@
 
             <EditText
                 android:id="@+id/et_text"
-                style="@style/Font.LogText"
+                style="@style/Font.SaveLogText"
                 android:layout_width="0dp"
                 android:layout_height="130dp"
                 android:layout_marginHorizontal="16dp"
@@ -118,7 +118,7 @@
 
             <Button
                 android:id="@+id/btn_save"
-                style="@style/Font.LogSave"
+                style="@style/Font.SaveLogSave"
                 android:layout_width="0dp"
                 android:layout_height="52dp"
                 android:layout_marginHorizontal="16dp"
