+++ app/src/main/res/layout/item_savelog.xml
@@ -28,7 +28,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:shapeAppearance="@style/LogImage" />
+            app:shapeAppearance="@style/SaveLogImage" />
 
         <ImageView
             android:id="@+id/iv_delete"
