+++ app/src/main/res/values/styles.xml
@@ -38,33 +38,33 @@
     </style>
 
     <!--기록 저장 화면 -->
-    <style name="LogTitle">
+    <style name="SaveLogTitle">
         <item name="android:fontFamily">@font/pretendard_semibold</item>
         <item name="android:textColor">@color/black</item>
         <item name="android:textSize">@dimen/large</item>
     </style>
 
-    <style name="LogPhoto">
+    <style name="SaveLogPhoto">
         <item name="android:fontFamily">@font/pretendard_semibold</item>
         <item name="android:textColor">@color/black</item>
         <item name="android:textSize">@dimen/middle</item>
     </style>
 
-    <style name="Font.LogText">
+    <style name="Font.SaveLogText">
         <item name="android:background">@drawable/button_record</item>
         <item name="android:gravity">top|start</item>
         <item name="android:textColor">@color/black</item>
         <item name="android:textSize">14sp</item>
     </style>
 
-    <style name="Font.LogSave">
+    <style name="Font.SaveLogSave">
         <item name="android:backgroundTint">@drawable/selector_button_record_background</item>
         <item name="android:textColor">@drawable/selector_button_record_text_color</item>
         <item name="android:textStyle">bold</item>
         <item name="android:textSize">@dimen/middle</item>
     </style>
 
-    <style name="LogImage">
+    <style name="SaveLogImage">
         <item name="cornerSize">12dp</item>
     </style>
 </resources>
